Extended ±999999999 year range for Date / parsing / duration math
Property::Date now stores days as i64, so the ±999999999-year range
the TCK exercises at its extremes round-trips. parse_iso_date_big
accepts the prefix-sign extended-year form (`-999999999-01-01`) and
big_ymd_to_days / big_days_to_ymd handle the calendar arithmetic
directly (bypassing chrono::NaiveDate, which tops out around
±262 K years).

duration.between / inMonths / inDays / inSeconds now operate on
integer day counts throughout, so a 2-billion-year range stays in
bounds of i64 days + i128 nanoseconds. duration_to_iso_string and
the TCK formatter widen their internal abs_ns / abs_secs split to
i128 so the 17_531_639_991_215-hour result prints correctly.

Temporal: 1002 → 1004 passing (100%). All 10 Temporal features at
100%.

+++ crates/mesh-core/src/property.rs
@@ -39,11 +39,11 @@ pub enum Property {
     /// Local (naive) datetime as epoch nanoseconds.
     /// Formatters omit any timezone suffix.
     LocalDateTime(i128),
-    /// Days since the UNIX epoch (1970-01-01, UTC). `i32` gives
-    /// ±5.9 million years of range — far more than any realistic
-    /// calendar application. Maps to Bolt `Date` (struct tag
-    /// `0x44`).
-    Date(i32),
+    /// Days since the UNIX epoch (1970-01-01, UTC). `i64` gives
+    /// roughly ±25 billion years — wide enough to round-trip the
+    /// ±999999999 year range the TCK exercises at its extremes.
+    /// Maps to Bolt `Date` (struct tag `0x44`), which is also i64.
+    Date(i64),
     /// A Cypher duration value — see [`Duration`].
     Duration(Duration),
     /// Time of day as nanoseconds since midnight, with an optional
